refactor: 컴포넌트 파일을 재 역할에 맞는 폴더로 이동
