+++ packages/web/src/app/_components/index.ts
@@ -1,5 +1,3 @@
-export { default as Introduction } from './Introduction';
 export { default as WebVitals } from './WebVitals';
 export { default as InitDayjs } from './InitDayjs';
-export { default as NotFoundSection } from './NotFoundSection';
 export { default as ErrorSection } from './ErrorSection';
