+++ packages/web/src/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './Config';
-export * from './Faq';
 export * from './Update';
